+++ src/untangled/client/data_fetch.cljs
@@ -15,15 +15,16 @@
   - `field`: A field on the current component's query that you wish to load
   - `without`: Named parameter for excluding child keys from the query (e.g. for recursive queries or additional laziness)
   - `params`: Named parameter for adding params to the query sent to the server for this field.
+  - `post-mutation`: A mutation (symbol) invoked after the load succeeds.
   "
-  [component field & {:keys [without params callback]}]
+  [component field & {:keys [without params post-mutation]}]
   (om/transact! component [(list 'app/load
                              {:ident    (om/get-ident component)
                               :field    field
                               :query    (om/focus-query (om/get-query component) [field])
                               :params   params
                               :without  without
-                              :callback callback})]))
+                              :post-mutation post-mutation})]))
 
 (defn load-collection
   "Load a collection from the remote. Runs `om/transact!`.
@@ -32,17 +33,18 @@
   - `comp-or-reconciler`: A component or reconciler (not a class)
   - `query`: The query for the element(s) attributes. Use defui to generate arbitrary queries so normalization will work.
   - Named parameter `ident`: An ident, used if loading a singleton and you wish to specify 'which one'.
+  - `post-mutation`: A mutation (symbol) invoked after the load succeeds.
 
   Named parameters `:without` and `:params` are as in `load-field`.
   "
-  [comp-or-reconciler query & {:keys [ident without params callback]}]
+  [comp-or-reconciler query & {:keys [ident without params post-mutation]}]
   (let []
     (om/transact! comp-or-reconciler [(list 'app/load
                                         {:ident    ident
                                          :query    query
                                          :params   params
                                          :without  without
-                                         :callback callback})])))
+                                         :post-mutation post-mutation})])))
 
 (def load-singleton load-collection)
 
@@ -57,7 +59,7 @@
     :action (fn []
        (load-field-action ...)
        ; other optimistic updates/state changes)}"
-  [app-state component-class ident field & {:keys [without params callback]}]
+  [app-state component-class ident field & {:keys [without params post-mutation]}]
   (impl/mark-ready
     :state app-state
     :field field
@@ -65,7 +67,7 @@
     :query (om/focus-query (om/get-query component-class) [field])
     :params params
     :without without
-    :callback callback))
+    :post-mutation post-mutation))
 
 (defn load-data-action
   "Queue up a remote load from within an already-running mutation. Similar to `load-collection`, but usable from
@@ -78,14 +80,14 @@
     :action (fn []
        (load-data-action ...)
        ; other optimistic updates/state changes)}"
-  [app-state query & {:keys [ident without params callback]}]
+  [app-state query & {:keys [ident without params post-mutation]}]
   (impl/mark-ready
     :state app-state
     :ident ident
     :query query
     :params params
     :without without
-    :callback callback))
+    :post-mutation post-mutation))
 
 
 (defn mark-loading