+++ src/untangled/client/impl/data_fetch.cljs
@@ -2,8 +2,7 @@
   (:require [om.next.impl.parser :as op]
             [om.next :as om]
             [clojure.walk :refer [prewalk]]
-            [cljs.core.async :as async]
-            )
+            [cljs.core.async :as async])
   (:require-macros
     [cljs.core.async.macros :refer [go]]))
 
@@ -77,7 +76,7 @@
 (defn ready-state
   "Generate a ready-to-load state with all of the necessary details to do
   remoting and merging."
-  [& {:keys [ident field params without query callback] :or {:without #{}}}]
+  [& {:keys [ident field params without query post-mutation] :or {:without #{}}}]
   (assert (or field query) "You must supply a query or a field/ident pair")
   (assert (or (not field) (and field (om/ident? ident))) "Field requires ident")
   (let [old-ast (om/query->ast query)
@@ -88,18 +87,18 @@
         key (if (om/join? query-field) (om/join-key query-field) query-field)
         query' (om/ast->query ast)]
     (assert (or (not field) (= field key)) "Component fetch query does not match supplied field.")
-    {::type     :ready
-     ::ident    ident                                       ; only for component-targeted loads
-     ::field    field                                       ; for component-targeted load
-     ::query    query'
-     ::callback callback}))                                 ; query, relative to root of db OR component
+    {::type          :ready
+     ::ident         ident                                  ; only for component-targeted loads
+     ::field         field                                  ; for component-targeted load
+     ::query         query'
+     ::post-mutation post-mutation}))                       ; query, relative to root of db OR component
 
 (defn mark-ready
   "Place a ready-to-load marker into the application state. This should be done from
   a mutate function that is abstractly loading something. This is intended for internal use.
 
   See `load-field` for public API."
-  [& {:keys [state query ident field without params callback] :or {:without #{}}}]
+  [& {:keys [state query ident field without params post-mutation] :or {:without #{}}}]
   (swap! state update :om.next/ready-to-load conj
     (ready-state
       :ident ident
@@ -107,7 +106,7 @@
       :params params
       :without without
       :query query
-      :callback callback)))
+      :post-mutation post-mutation)))
 
 ;; TODO: Rename "getters"
 (defn data-ident [state] (::ident state))
@@ -184,7 +183,11 @@
           app-state (om/app-state reconciler)]
 
       (om/merge! reconciler response query)
-      (doseq [item loading-items] (if-let [cb (::callback item)] (cb (om/app-state reconciler))))
+      (doseq [item loading-items] (when-let [mutation-symbol (::post-mutation item)]
+                                    (some->
+                                      (untangled.client.mutations/mutate {:state (om/app-state reconciler)} mutation-symbol {})
+                                      :action
+                                      (apply []))))
       (om/force-root-render! reconciler)                    ; Don't love this, but ok for now. TK
 
       ;; Any loading states that didn't resolve to data are marked as not present